Consolidate Discord emoji helpers
